+++ app/src/main/java/com/intencity/intencity/helper/doa/ExerciseDao.java
@@ -60,13 +60,13 @@ public class ExerciseDao
                 String notes = object.getString(Constant.COLUMN_NOTES);
 
                 Set set = new Set();
-                set.setWeight(weight.equalsIgnoreCase(Constant.RETURN_NULL) ? Constant.CODE_FAILED :
+                set.setWeight(weight.equalsIgnoreCase(Constant.RETURN_NULL) ? (int) Constant.CODE_FAILED :
                                       Float.valueOf(weight));
                 set.setReps(reps.equalsIgnoreCase(Constant.RETURN_NULL) ? 0 :
                                     Integer.valueOf(reps));
                 set.setDuration(duration);
                 set.setDifficulty(difficulty.equalsIgnoreCase(Constant.RETURN_NULL) ?
-                                          Constant.CODE_FAILED : Integer.valueOf(difficulty));
+                                          (int) Constant.CODE_FAILED : Integer.valueOf(difficulty));
                 set.setNotes(notes);
 
                 ArrayList<Set> sets = new ArrayList<>();
